requirements.txt
numpy>=1.21.0
torch>=1.9.0
gymnasium>=0.26.0
minigrid>=2.0.0
PyYAML>=6.0
# Optional: comet-ml  (only needed if logger.type == "comet")
